Changed particles from a soup of electrons to particles under the force of gravity. Other quasi-forces may be considered, although, gravity is only attractive, there are no other actual forces with this process.
The electron is a subatomic particle, symbol e− or β−, with a negative elementary electric charge. Electrons belong to the first generation of the lepton particle family, and are generally thought to be elementary particles because they have no known components or substructure. The electron has a mass that is approximately 1/1836 that of the proton. Quantum mechanical properties of the electron include an intrinsic angular momentum (spin) of a half-integer value in units of ħ, which means that it is a fermion. Being fermions, no two electrons can occupy the same quantum state, in accordance with the Pauli exclusion principle. Like all matter, electrons have properties of both particles and waves, and so can collide with other particles and can be diffracted like light. The wave properties of electrons are easier to observe with experiments than those of other particles like neutrons and protons because electrons have a lower mass and hence a higher De Broglie wavelength for typical energies.
Many physical phenomena involve electrons in an essential role, such as electricity, magnetism, and thermal conductivity, and they also participate in gravitational, electromagnetic and weak interactions. An electron generates an electric field surrounding it. An electron moving relative to an observer generates a magnetic field. External magnetic fields deflect an electron. Electrons radiate or absorb energy in the form of photons when accelerated. Laboratory instruments are capable of containing and observing individual electrons as well as electron plasma using electromagnetic fields, whereas dedicated telescopes can detect electron plasma in outer space. Electrons have many applications, including electronics, welding, cathode ray tubes, electron microscopes, radiation therapy, lasers, gaseous ionization detectors and particle accelerators.
Interactions involving electrons and other subatomic particles are of interest in fields such as chemistry and nuclear physics. The Coulomb force interaction between positive protons inside atomic nuclei and negative electrons composes atoms. Ionization or changes in the proportions of particles changes the binding energy of the system. The exchange or sharing of the electrons between two or more atoms is the main cause of chemical bonding. British natural philosopher Richard Laming first hypothesized the concept of an indivisible quantity of electric charge to explain the chemical properties of atoms in 1838; Irish physicist George Johnstone Stoney named this charge 'electron' in 1891, and J. J. Thomson and his team of British physicists identified it as a particle in 1897. Electrons can also participate in nuclear reactions, such as nucleosynthesis in stars, where they are known as beta particles. Electrons may be created through beta decay of radioactive isotopes and in high-energy collisions, for instance when cosmic rays enter the atmosphere. The antiparticle of the electron is called the positron; it is identical to the electron except that it carries electrical and other charges of the opposite sign. When an electron collides with a positron, both particles may be totally annihilated, producing gamma ray photons.

Gravity or gravitation is a natural phenomenon by which all things with mass are brought towards (or 'gravitate' towards) one another including stars, planets, galaxies and even light and sub-atomic particles. Gravity is responsible for the complexity in the universe, by creating spheres of hydrogen, igniting them under pressure to form stars and grouping them into galaxies. Without gravity, the universe would be an uncomplicated one, existing without thermal energy and composed only of equally spaced particles. On Earth, gravity gives weight to physical objects and causes the tides. Gravity has an infinite range, and it cannot be absorbed, transformed, or shielded against.
Gravity is most accurately described by the general theory of relativity (proposed by Albert Einstein in 1915) which describes gravity, not as a force, but as a consequence of the curvature of spacetime caused by the uneven distribution of mass/energy; and resulting in time dilation, where time lapses more slowly in strong gravitation. However, for most applications, gravity is well approximated by Newton's law of universal gravitation, which postulates that gravity is a force where two bodies of mass are directly drawn (or 'attracted') to each other according to a mathematical relationship, where the attractive force is proportional to the product of their masses and inversely proportional to the square of the distance between them. This is considered to occur over an infinite range, such that all bodies (with mass) in the universe are drawn to each other no matter how far they are apart.
Gravity is the weakest of the four fundamental interactions of nature. The gravitational attraction is approximately 10−38 times the strength of the strong force (i.e. gravity is 38 orders of magnitude weaker), 10−36 times the strength of the electromagnetic force, and 10−29 times the strength of the weak force. As a consequence, gravity has a negligible influence on the behavior of sub-atomic particles, and plays no role in determining the internal properties of everyday matter (but see quantum gravity). On the other hand, gravity is the dominant force at the macroscopic scale, that is the cause of the formation, shape, and trajectory (orbit) of astronomical bodies, including those of asteroids, comets, planets, stars, and galaxies. It is responsible for causing the Earth and the other planets to orbit the Sun; for causing the Moon to orbit the Earth; for the formation of tides; for natural convection, by which fluid flow occurs under the influence of a density gradient and gravity; for heating the interiors of forming stars and planets to very high temperatures; for solar system, galaxy, stellar formation and evolution; and for various other phenomena observed on Earth and throughout the universe.
In pursuit of a theory of everything, the merging of general relativity and quantum mechanics (or quantum field theory) into a more general theory of quantum gravity has become an area of research.